remove Compactable, Filterable, TraversableFilterable and /data folder
also remove SemiCoproduct/coproductEither
