+++ .changeset/swift-melons-reflect.md
@@ -1,0 +1,5 @@
+---
+"@fp-ts/core": patch
+---
+
+remove Compactable, Filterable, TraversableFilterable and /data folder
