--- src/data/Option.ts
@@ -1,35 +1,0 @@
-/**
- * @since 1.0.0
- */
-import type { TypeLambda } from "@fp-ts/core/HKT"
-
-/**
- * @category models
- * @since 1.0.0
- */
-export interface None {
-  readonly _tag: "None"
-}
-
-/**
- * @category models
- * @since 1.0.0
- */
-export interface Some<out A> {
-  readonly _tag: "Some"
-  readonly value: A
-}
-
-/**
- * @category models
- * @since 1.0.0
- */
-export type Option<A> = None | Some<A>
-
-/**
- * @category type lambdas
- * @since 1.0.0
- */
-export interface OptionTypeLambda extends TypeLambda {
-  readonly type: Option<this["Target"]>
-}
