--- src/internal/Option.ts
@@ -1,24 +1,0 @@
-/**
- * @since 1.0.0
- */
-import type { None, Option, Some } from "@fp-ts/core/data/Option"
-
-/**
- * @since 1.0.0
- */
-export const none: Option<never> = { _tag: "None" }
-
-/**
- * @since 1.0.0
- */
-export const some = <A>(a: A): Option<A> => ({ _tag: "Some", value: a })
-
-/**
- * @since 1.0.0
- */
-export const isNone = <A>(fa: Option<A>): fa is None => fa._tag === "None"
-
-/**
- * @since 1.0.0
- */
-export const isSome = <A>(fa: Option<A>): fa is Some<A> => fa._tag === "Some"
