+++ test/data/NonEmptyReadonlyArray.ts
