+++ test/data/Predicate.ts
