+++ test/data/number.ts
