+++ test/data/string.ts
