+++ test/typeclass/Of.ts
@@ -1,6 +1,6 @@
 import * as _ from "@fp-ts/core/typeclass/Of"
-import * as O from "../test-data/Option"
-import * as RA from "../test-data/ReadonlyArray"
+import * as O from "../data/Option"
+import * as RA from "../data/ReadonlyArray"
 import * as U from "../util"
 
 describe("Of", () => {
