+++ test/typeclass/Product.ts
@@ -1,10 +1,10 @@
 import { pipe } from "@fp-ts/core/internal/Function"
 import * as _ from "@fp-ts/core/typeclass/Product"
 import * as semigroup from "@fp-ts/core/typeclass/Semigroup"
-import * as number from "../test-data/number"
-import * as O from "../test-data/Option"
-import * as P from "../test-data/Predicate"
-import * as string from "../test-data/string"
+import * as number from "../data/number"
+import * as O from "../data/Option"
+import * as P from "../data/Predicate"
+import * as string from "../data/string"
 import * as U from "../util"
 
 describe("Product", () => {
