+++ test/typeclass/SemiApplicative.ts
@@ -1,7 +1,7 @@
 import { pipe } from "@fp-ts/core/internal/Function"
 import * as _ from "@fp-ts/core/typeclass/SemiApplicative"
-import * as O from "../test-data/Option"
-import * as string from "../test-data/string"
+import * as O from "../data/Option"
+import * as string from "../data/string"
 import * as U from "../util"
 
 describe("SemiApplicative", () => {
