+++ test/typeclass/SemiProduct.ts
@@ -3,11 +3,11 @@ import { pipe } from "@fp-ts/core/internal/Function"
 import * as semiApplicative from "@fp-ts/core/typeclass/SemiApplicative"
 import * as semigroup from "@fp-ts/core/typeclass/Semigroup"
 import * as _ from "@fp-ts/core/typeclass/SemiProduct"
-import * as number from "../test-data/number"
-import * as O from "../test-data/Option"
-import * as P from "../test-data/Predicate"
-import * as RA from "../test-data/ReadonlyArray"
-import * as string from "../test-data/string"
+import * as number from "../data/number"
+import * as O from "../data/Option"
+import * as P from "../data/Predicate"
+import * as RA from "../data/ReadonlyArray"
+import * as string from "../data/string"
 import * as U from "../util"
 
 describe("SemiProduct", () => {
