+++ test/typeclass/Semigroup.ts
@@ -1,8 +1,8 @@
 import { pipe } from "@fp-ts/core/internal/Function"
 import * as order from "@fp-ts/core/typeclass/Order"
 import * as _ from "@fp-ts/core/typeclass/Semigroup"
-import * as number from "../test-data/number"
-import * as string from "../test-data/string"
+import * as number from "../data/number"
+import * as string from "../data/string"
 import * as U from "../util"
 
 describe("Semigroup", () => {
