+++ test/typeclass/Traversable.ts
@@ -1,7 +1,7 @@
 import { pipe } from "@fp-ts/core/internal/Function"
 import * as _ from "@fp-ts/core/typeclass/Traversable"
-import * as O from "../test-data/Option"
-import * as RA from "../test-data/ReadonlyArray"
+import * as O from "../data/Option"
+import * as RA from "../data/ReadonlyArray"
 import * as U from "../util"
 
 describe("Traversable", () => {
