+++ test/typeclass/TraversableWithIndex.ts
@@ -1,7 +1,7 @@
 import { pipe } from "@fp-ts/core/internal/Function"
+import * as O from "../data/Option"
+import * as RA from "../data/ReadonlyArray"
 import * as _ from "../limbo/TraversableWithIndex"
-import * as O from "../test-data/Option"
-import * as RA from "../test-data/ReadonlyArray"
 import * as U from "../util"
 
 describe("TraversableWithIndex", () => {
