+++ vitest.config.ts
@@ -7,7 +7,7 @@ export default defineConfig({
     include: ["./test/**/*.{js,mjs,cjs,ts,mts,cts,jsx,tsx}"],
     exclude: [
       "./test/util.ts",
-      "./test/test-data/*.ts",
+      "./test/data/*.ts",
       "./test/limbo/*.ts"
     ],
     globals: true
